Revert to default sphinx theme
Remove dependency on EllisLab corporate theme - that is not appropriate for us.
Reverted to default Sphinx theme, with a couple of options, which have the side-effect of adding back document tree navigation (next/previous) and collapsible menu (albeit a sidebar not a top one).

Closes #2605 and closes #2354, for now.
This is not a proper styling - that is a separate issue!
Signed-off-by:James L Parry <jim_parry@bcit.ca>