Add pulldown menu to user guide.
Integrating Rufnex's Javascript optional menu replacement :)

Now you will see a small icon in the top right of the user guide browser window. Clicking this toggles between the readdocs style TOC, in the left sidebar, and the "classic" CI TOC, a pulldown activated by clicking on the "Table of Contents" icon at the top of the browser window. The tool use a cookie to remember your decision.

Integrated on behalf of .... -- Have fun! Rufnex - codeigniter.de